[Synthetics] Extract buildMonitorParamsSearch helper for breadcrumb/link URLs
