+++ x-pack/solutions/observability/plugins/synthetics/public/apps/synthetics/components/test_run_details/hooks/use_test_run_details_breadcrumbs.ts
@@ -11,6 +11,7 @@ import { useSelectedMonitor } from '../../monitor_details/hooks/use_selected_mon
 import { useBreadcrumbs } from '../../../hooks/use_breadcrumbs';
 import { useUrlSpaceId } from '../../../hooks/use_url_space_id';
 import { useGetUrlParams } from '../../../hooks';
+import { buildMonitorParamsSearch } from '../../../utils/url_params';
 import { ConfigKey } from '../../../../../../common/runtime_types';
 import { MONITOR_ROUTE, MONITORS_ROUTE } from '../../../../../../common/constants';
 import { PLUGIN } from '../../../../../../common/constants/plugin';
@@ -28,11 +29,11 @@ export const useTestRunDetailsBreadcrumbs = (
   // doesn't drop into the local saved-object 404 page.
   const { remoteName } = useGetUrlParams();
 
-  const monitorHrefParams = new URLSearchParams();
-  if (selectedLocation?.id) monitorHrefParams.set('locationId', selectedLocation.id);
-  if (spaceId) monitorHrefParams.set('spaceId', spaceId);
-  if (remoteName) monitorHrefParams.set('remoteName', remoteName);
-  const monitorHrefSearch = monitorHrefParams.toString();
+  const monitorSearch = buildMonitorParamsSearch({
+    locationId: selectedLocation?.id,
+    spaceId,
+    remoteName,
+  });
 
   useBreadcrumbs([
     {
@@ -46,7 +47,7 @@ export const useTestRunDetailsBreadcrumbs = (
             href: `${appPath}${MONITOR_ROUTE.replace(
               ':monitorId',
               monitor?.[ConfigKey.CONFIG_ID] ?? ''
-            )}${monitorHrefSearch ? `?${monitorHrefSearch}` : ''}`,
+            )}${monitorSearch}`,
           },
         ]
       : []),
